Updated Game.cpp and Player.cpp
I updated the comments of some of the methods and added some cout statements for the user to know a little better what is happening.

I added player as one of the parameters of split() so that the second hand can be replaced. However, I'm having trouble replacing the second hand with the first hand. (Added a setHand in Player). Maybe it has something to do with pointers again!