Updated main.cpp
Declared p1 as the first account.

Moved comments around from instructions to match the correct spots of where we implemented.

Added more cout statments to allow the user to know what is happening during the game.

Updated bet amounts to update at the time of deciding if they win.

FIX: The split() function does not update the hand correctly yet.